Make it AOO frendly and add some features to the app

+++ src/main/java/l3/miage/shopcook/services/IngredientServiceImpl.java
@@ -8,45 +8,46 @@ import org.springframework.stereotype.Service;
 import l3.miage.shopcook.ingredients.Ingredient;
 import l3.miage.shopcook.repositories.IngredientRepository;
 import l3.miage.shopcook.user.User;
+
 @Service
 public class IngredientServiceImpl implements IngredientService {
     @Autowired
     private IngredientRepository ingredientRepository;
+
     @Override
-    public boolean save(Ingredient ingredient) {
+    public Ingredient save(Ingredient ingredient) {
         try {
- 
-            this.ingredientRepository.save(ingredient);
-            return true;
-        }
-        catch (Exception e) {
-            return false;
+            return this.ingredientRepository.save(ingredient);
+        } catch (Exception e) {
+
+            return null;
         }
     }
-    public boolean save (String name, float nutriscore, int calories) {
+
+    public boolean save(String name, float nutriscore, int calories) {
         Ingredient i = new Ingredient(name, nutriscore, calories);
         this.save(i);
         return true;
     }
+
     @Override
-    public boolean removeIngredient(Ingredient ingredient) {
+    public Ingredient removeIngredient(Ingredient ingredient) {
         try {
             this.ingredientRepository.delete(ingredient);
-            return true;
-        }
-        catch (Exception e) {
-            return false;
+            return ingredient;
+        } catch (Exception e) {
+
+            return null;
         }
     }
 
     @Override
-    public boolean updateIngredient(Ingredient ingredient) {
+    public Ingredient updateIngredient(Ingredient ingredient) {
         if (this.ingredientRepository.existsById(ingredient.getId())) {
             this.ingredientRepository.deleteById(ingredient.getId());
-            this.save(ingredient);
-            return true;
+            return this.save(ingredient);
         }
-        return false;
+        return null;
     }
 
     @Override
@@ -58,13 +59,19 @@ public class IngredientServiceImpl implements IngredientService {
     public List<Ingredient> findAll() {
         return this.ingredientRepository.findAll();
     }
+
     @Override
-    public List<Ingredient> findByUser(User user) { 
-       return this.ingredientRepository.findByUser(user);
+    public List<Ingredient> findByUser(User user) {
+        return this.ingredientRepository.findByUser(user);
     }
+
     @Override
-    public void deleteById(Integer id) {
-         this.ingredientRepository.deleteById(id);
+    public Ingredient deleteById(Integer id) {
+        Ingredient ingredient = this.ingredientRepository.findById(id).orElse(null);
+        if (ingredient != null) {
+            this.ingredientRepository.deleteById(id);
+        }
+        return ingredient;
     }
-    
+
 }
